Deleted unnecessary userDocumentControllerTest and centralized tests
