+++ itachallenge-user/src/test/java/com/itachallenge/user/controller/UserControllerTest.java
@@ -1,10 +1,7 @@
 package com.itachallenge.user.controller;
 
 import com.itachallenge.user.document.UserSolutionDocument;
-import com.itachallenge.user.dtos.BookmarkRequestDto;
-import com.itachallenge.user.dtos.ChallengeStatisticsDto;
-import com.itachallenge.user.dtos.SolutionUserDto;
-import com.itachallenge.user.dtos.UserScoreDto;
+import com.itachallenge.user.dtos.*;
 import com.itachallenge.user.service.IServiceChallengeStatistics;
 import com.itachallenge.user.service.IUserSolutionService;
 import org.junit.jupiter.api.BeforeEach;
@@ -22,6 +19,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import reactor.core.publisher.Mono;
 
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
@@ -29,7 +27,7 @@ import java.util.UUID;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -209,6 +207,102 @@ class UserControllerTest {
         assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
 
         assertEquals(bookmarkRequestDto, responseEntity.getBody());
+    }
+
+    @Test
+    void addSolution_ValidSolution_SolutionAdded() {
+        String URI_TEST = "/solution";
+        UserSolutionDto userSolutionDto = new UserSolutionDto();
+        userSolutionDto.setUserId("550e8400-e29b-41d4-a716-446655440001");
+        userSolutionDto.setChallengeId("550e8400-e29b-41d4-a716-446655440002");
+        userSolutionDto.setLanguageId("550e8400-e29b-41d4-a716-446655440003");
+        userSolutionDto.setSolutionText("This is a test solution");
+
+        UserSolutionScoreDto expectedResponse = new UserSolutionScoreDto(userSolutionDto.getUserId(),
+                userSolutionDto.getChallengeId(), userSolutionDto.getLanguageId(),
+                userSolutionDto.getSolutionText(), 13);
+
+        when(userScoreService.addSolution(userSolutionDto.getUserId(),
+                userSolutionDto.getChallengeId(),
+                userSolutionDto.getLanguageId(),
+                userSolutionDto.getSolutionText()))
+                .thenReturn(Mono.just(expectedResponse));
+
+        webTestClient.put()
+                .uri(CONTROLLER_URL + URI_TEST)
+                .contentType(MediaType.APPLICATION_JSON)
+                .bodyValue(userSolutionDto)
+                .exchange()
+                .expectStatus().isEqualTo(HttpStatus.ACCEPTED)
+                .expectBody(UserSolutionScoreDto.class)
+                .value(dto -> {
+                    assert dto != null;
+                    assert dto.getUserId() != null;
+                    assert dto.getChallengeId() != null;
+                    assert dto.getLanguageId() != null;
+                    assert dto.getScore() >= 0;
+
+                    verify(userScoreService, times(1)).addSolution(
+                            userSolutionDto.getUserId(),
+                            userSolutionDto.getChallengeId(),
+                            userSolutionDto.getLanguageId(),
+                            userSolutionDto.getSolutionText()
+                    );
+                });
+    }
+
+    @Test
+    void addSolution_InvalidValues_BadRequest() {
+        String URI_TEST = "/solution";
+
+        List<UserSolutionDto> testCases = Arrays.asList(
+                new UserSolutionDto("invalid_uuid", "550e8400-e29b-41d4-a716-446655440002", "550e8400-e29b-41d4-a716-446655440003", "This is a test solution"),
+                new UserSolutionDto("550e8400-e29b-41d4-a716-446655440001", "invalid_uuid", "550e8400-e29b-41d4-a716-446655440003", "This is a test solution"),
+                new UserSolutionDto("550e8400-e29b-41d4-a716-446655440001", "550e8400-e29b-41d4-a716-446655440002", "invalid_uuid", "This is a test solution"),
+                new UserSolutionDto("550e8400-e29b-41d4-a716-446655440001", "550e8400-e29b-41d4-a716-446655440002", "550e8400-e29b-41d4-a716-446655440003", "")
+        );
+
+        for (UserSolutionDto testCase : testCases) {
+            webTestClient.put()
+                    .uri(CONTROLLER_URL + URI_TEST)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .bodyValue(testCase)
+                    .exchange()
+                    .expectStatus().isBadRequest();
+        }
+    }
+
+    @Test
+    void addSolution_EmptyRequestBody_BadRequest() {
+        String URI_TEST = "/solution";
+        UserSolutionDto userSolutionDto = new UserSolutionDto();
+
+        webTestClient.put()
+                .uri(CONTROLLER_URL + URI_TEST)
+                .contentType(MediaType.APPLICATION_JSON)
+                .bodyValue(userSolutionDto)
+                .exchange()
+                .expectStatus().isBadRequest();
+    }
+
+    @Test
+    void addSolution_ServiceThrowsException_InternalServerError() {
+        String URI_TEST = "/solution";
+        UserSolutionDto userSolutionDto = new UserSolutionDto();
+        userSolutionDto.setUserId("550e8400-e29b-41d4-a716-446655440001");
+        userSolutionDto.setChallengeId("550e8400-e29b-41d4-a716-446655440002");
+        userSolutionDto.setLanguageId("550e8400-e29b-41d4-a716-446655440003");
+        userSolutionDto.setSolutionText("This is a test solution");
+
+        when(userScoreService.addSolution(any(), any(), any(), any()))
+                .thenReturn(Mono.error(new RuntimeException("Test exception")));
+
+        webTestClient.put()
+                .uri(CONTROLLER_URL + URI_TEST)
+                .contentType(MediaType.APPLICATION_JSON)
+                .bodyValue(userSolutionDto)
+                .exchange()
+                .expectStatus().is5xxServerError();
     }
